+++ src/main.rs
@@ -1,0 +1,10 @@
+
+#[allow(dead_code)]
+
+mod rlp;
+
+
+#[cfg(not(test))]
+fn main() {
+	println!("Ethereum");
+}
